Refactor MultipleLocationInput Component: Separate Concerns
Refactored the MultipleLocationInput component to separate concerns and enhance code maintainability:

- **/hooks/useSearch.ts**: Created a custom hook to handle the search functionality of the component. This allowed for the removal of search logic from the component itself and imported the useSearch hook.
- **/schema/MultipleLocationInputProps.ts**: Exported the props that were declared in the component to a separate file as a best practice.

This restructuring makes the code more modular and easier to understand, following best practices for organizing React components. By separating the search logic and props into their own files, the codebase becomes more maintainable and scalable.
