+++ CtasApp/src/components/MultipleLocationInput.tsx
@@ -1,58 +1,31 @@
 import * as React from 'react';
-import {useState, useEffect} from 'react';
+import {useState} from 'react';
 import {View, TextInput, FlatList, TouchableOpacity, Text} from 'react-native';
 import InputFieldStyles from '../styles/InputFieldStyles';
 import Icon from 'react-native-vector-icons/AntDesign';
 import Plus from 'react-native-vector-icons/FontAwesome6';
 import MessageModal from './MessageModal';
-import { truncateString } from '../helpers/TruncateStringHelper';
-interface MultipleLocationInputProps {
-  onSelected?: (items: string[]) => void;
-  errorMessage?: string;
-}
+import {truncateString} from '../helpers/TruncateStringHelper';
+import {MultipleLocationInputProps} from '../schema/MultipleLocationInputProps';
+import useSearch from '../hooks/useSearch';
 
 const MultipleLocationInput: React.FC<MultipleLocationInputProps> = ({
   onSelected,
   errorMessage,
 }) => {
   const [query, setQuery] = useState<string>('');
-  const [results, setResults] = useState<string[]>([]);
-  const [showResults, setShowResults] = useState<boolean>(false);
   const [selectedItems, setSelectedItems] = useState<string[]>([]);
   const [showAddresses, setShowAddresses] = useState<boolean>(false);
   const [isModalVisible, setIsModalVisible] = useState<boolean>(false);
   const [isInputDisabled, setIsInputDisabled] = useState<boolean>(false);
 
-  useEffect(() => {
-    if (query.trim() === '') {
-      setResults([]);
-      setShowResults(false);
-      return;
-    }
-
-    const handleSearch = () => {
-      fetch(
-        `https://api.geoapify.com/v1/geocode/autocomplete?text=${query}&format=json&apiKey=6c07146c0b5f4db080b4f024e2624690`,
-      )
-        .then(response => response.json())
-        .then(result => {
-          const processedResults = result.results
-            ? result.results.map((item: {formatted: any}) => item.formatted)
-            : [];
-          setResults(processedResults);
-          setShowResults(processedResults.length > 0);
-        })
-        .catch(error => console.log('error', error));
-    };
-
-    handleSearch();
-  }, [query]);
+  const {results, showResults, setShowResults, setResults} = useSearch(query);
 
   const handleSelectItem = (item: string) => {
     setSelectedItems([...selectedItems, item]);
     setResults([]);
-    setShowResults(false); 
-    setQuery('Añade + direcciones desplegando el menú "v"');
+    setShowResults(false);
+    setQuery('Añade + direcciones desplegando el menú');
     setIsInputDisabled(true);
     onSelected?.([...selectedItems, item]);
     setIsModalVisible(true);
@@ -60,7 +33,6 @@ const MultipleLocationInput: React.FC<MultipleLocationInputProps> = ({
       setIsModalVisible(false);
     }, 1500);
   };
-  
 
   const handleAddNewAddress = () => {
     setQuery('');
@@ -113,7 +85,9 @@ const MultipleLocationInput: React.FC<MultipleLocationInputProps> = ({
           {selectedItems.map((item, index) => (
             <View key={index}>
               <View style={InputFieldStyles.selectedItemContainer}>
-                <Text style={InputFieldStyles.selectedItemText}>{truncateString(item, 22)}</Text>
+                <Text style={InputFieldStyles.selectedItemText}>
+                  {truncateString(item, 22)}
+                </Text>
                 <TouchableOpacity
                   onPress={() => handleDeleteAddress(index)}
                   style={InputFieldStyles.trashContainer}>
