+++ CtasApp/src/schema/MultipleLocationInputProps.ts
@@ -1,0 +1,7 @@
+interface MultipleLocationInputProps {
+    onSelected?: (items: string[]) => void;
+    errorMessage?: string;
+  }
+  
+  export { MultipleLocationInputProps };
+
